pyproject.toml
[tool.poetry]
name = "flaskbot"
version = "0.1.0"
description = ""
authors = ["Artem <106368820+A-V-tor@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
flask-sqlalchemy = "2.5.1"
pyTelegramBotAPI = "^4.7.1"
python-dotenv = "^0.21.0"
Flask-Admin = "^1.6.0"
Pillow = "^9.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
